+++ core-ui/src/shared/ResourceForm/components/FormField.js
@@ -1,0 +1,33 @@
+import React from 'react';
+import classnames from 'classnames';
+
+import { Label } from './Label';
+import './FormComponents.scss';
+
+export function FormField({
+  simple,
+  advanced,
+  propertyPath,
+  label,
+  input,
+  className,
+  required,
+  disabled,
+  tooltipContent,
+  isAdvanced,
+  defaultValue,
+  ...props
+}) {
+  return (
+    <div className={classnames('fd-row form-field', className)}>
+      <div className="fd-col fd-col-md--4 form-field__label">
+        <Label required={required && !disabled} tooltipContent={tooltipContent}>
+          {label}
+        </Label>
+      </div>
+      <div className="fd-col fd-col-md--7">
+        {input({ required, disabled, ...props })}
+      </div>
+    </div>
+  );
+}
