+++ core-ui/src/shared/ResourceForm/components/ResourceForm.scss
