+++ core-ui/src/shared/ResourceForm/inputs/Switch.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Switch as FundamentalSwitch } from 'fundamental-react';
+
+export function Switch({ value, setValue, ...props }) {
+  return (
+    <FundamentalSwitch
+      compact
+      onChange={e => setValue(!value)}
+      checked={value}
+      {...props}
+    />
+  );
+}
