+++ core-ui/src/shared/ResourceForm/inputs/Text.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { FormInput } from 'fundamental-react';
+
+export function Text({ value, setValue, ...props }) {
+  return (
+    <FormInput
+      compact
+      value={value}
+      onChange={e => setValue(e.target.value)}
+      {...props}
+    />
+  );
+}
