+++ core-ui/src/shared/ResourceForm/inputs/index.js
@@ -1,0 +1,5 @@
+export * from './Text';
+export * from './Number';
+export * from './Switch';
+export * from './Dropdown';
+export * from '.Checkboxes';
